fix: 修复空指针问题
修复一些空指针的问题

+++ src/main/java/org/fatmansoft/teach/dto/TotalRankDTO.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class TotalRankDTO {
     private Integer rank;
     private Double percent;
-    private Integer averageScore;
+    private Double averageScore;
     private Integer sameScoreNum;
 
     @Override
